requirements.txt
numpy>=1.21.0
torch>=1.10.0
gymnasium>=0.28.1
matplotlib>=3.5.0
tensorboard>=2.8.0
PyOpenGL>=3.1.6
pygame>=2.1.0
imageio>=2.19.0
Pillow>=9.0.0
tqdm>=4.62.0
pandas>=1.3.0
seaborn>=0.11.2